[refactor] Move feature and feature config into LDFeatureImporter class
